add new homebottommenuscreen
