+++ lib/bottom_works_with_riverpod/hide_show_bottom_nav_done.dart
@@ -3,6 +3,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../Constant/constants.dart';
+import '../Screens/home_bottom_menu.dart';
 import '../Screens/second_screen.dart';
 import '../provider.dart';
 
@@ -48,13 +49,13 @@ class MainScreen extends ConsumerWidget {
                 child: IndexedStack(
                   index: indexBottomNavbar,
                   children: [
+                    const HomeBottomMenuScreen(),
                     Center(
                       child: ElevatedButton(
                         onPressed: () => ref.read(goRouterProvider).go('/details'),
                         child: const Text('Go to the Details screen'),
                       ),
                     ),
-                    const HomeScreen(),
                     const Center(
                       child: Text("Profile"),
                     ),
@@ -99,114 +100,4 @@ class MainScreen extends ConsumerWidget {
 
 }
 
-class HomeScreen extends ConsumerStatefulWidget  {
 
-  const HomeScreen({Key? key,
-  }) : super(key: key);
-
-  @override
-  ConsumerState<HomeScreen> createState() => _HomeScreenState();
-}
-
-class _HomeScreenState extends ConsumerState<HomeScreen> {
-  ScrollController scrollController = ScrollController();
-
-  @override
-  void initState() {
-    super.initState();
-
-    scrollController.addListener(() {
-      if(scrollController.position.userScrollDirection == ScrollDirection.forward) {
-        ref.read(visibilityProvider.notifier).state = true;
-      } else {
-        ref.read(visibilityProvider.notifier).state = false;
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-
-    scrollController.removeListener(() {
-      if(scrollController.position.userScrollDirection == ScrollDirection.forward) {
-        ref.read(visibilityProvider.notifier).state = true;
-      } else {
-        ref.read(visibilityProvider.notifier).state = false;
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView(
-      controller: scrollController,
-      children: [
-        Container(
-          color: Colors.red,
-          height: 270,
-        ),
-        Container(
-          color: Colors.blue,
-          height: 200,
-        ),
-        Container(
-          color: Colors.purple,
-          height: 200,
-        ),
-        Container(
-          color: Colors.green,
-          height: 280,
-        ),
-        Container(
-          color: Colors.red,
-          height: 270,
-        ),
-        Container(
-          color: Colors.blue,
-          height: 200,
-        ),
-        Container(
-          color: Colors.purple,
-          height: 200,
-        ),
-        Container(
-          color: Colors.green,
-          height: 280,
-        ),
-        Container(
-          color: Colors.red,
-          height: 270,
-        ),
-        Container(
-          color: Colors.blue,
-          height: 200,
-        ),
-        Container(
-          color: Colors.purple,
-          height: 200,
-        ),
-        Container(
-          color: Colors.green,
-          height: 280,
-        ),
-        Container(
-          color: Colors.red,
-          height: 270,
-        ),
-        Container(
-          color: Colors.blue,
-          height: 200,
-        ),
-        Container(
-          color: Colors.purple,
-          height: 200,
-        ),
-        Container(
-          color: Colors.green,
-          height: 280,
-        ),
-      ],
-    );
-  }
-}
